+++ apps/nextjs-app/src/features/home/home.config.ts
@@ -1,10 +1,11 @@
 import type { I18nActiveNamespaces } from '@/lib/i18n';
 
-export type HomeConfig = {
+export interface IHomeConfig {
   // Define namespaces in use in both the type and the config.
   i18nNamespaces: I18nActiveNamespaces<'common' | 'home'>;
-};
-export const homeConfig: HomeConfig = {
+}
+
+export const homeConfig: IHomeConfig = {
   /** Namespaces that should be loaded for this page */
   i18nNamespaces: ['common', 'home'],
 };