+++ apps/nextjs-app/src/types.d/react-i18next.d.ts
@@ -8,6 +8,7 @@ import 'react-i18next';
 import type { I18nNamespaces } from '@teable-group/common-i18n';
 
 declare module 'react-i18next' {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
   interface CustomTypeOptions {
     defaultNS: 'common';
     resources: I18nNamespaces;