+++ packages/sdk/src/api/json-api/json-api.typeguard.ts
@@ -1,25 +1,25 @@
 import { isPlainObject } from '@teable-group/ts-utils';
 import type {
-  JsonApiErrorResponse,
-  JsonApiResponse,
-  JsonApiSuccessResponse,
+  IJsonApiErrorResponse,
+  IJsonApiResponse,
+  IJsonApiSuccessResponse,
 } from './json-api-response.types';
 
 export const isJsonApiResponse = <T = unknown>(
   val: unknown
-): val is JsonApiResponse<T> => {
+): val is IJsonApiResponse<T> => {
   return isPlainObject(val) && typeof val?.success === 'boolean';
 };
 
 export const isJsonApiSuccessResponse = <T = unknown>(
   val: unknown
-): val is JsonApiSuccessResponse<T> => {
+): val is IJsonApiSuccessResponse<T> => {
   return isJsonApiResponse<T>(val) && val.success && 'data' in val;
 };
 
 export const isJsonApiErrorResponse = (
   val: unknown
-): val is JsonApiErrorResponse => {
+): val is IJsonApiErrorResponse => {
   return (
     isJsonApiResponse<unknown>(val) &&
     !val.success &&