+++ packages/sdk/src/hooks/use-deep-compare-memoize.ts
@@ -1,15 +1,15 @@
 import { dequal } from 'dequal';
 import { useRef } from 'react';
 
-type SupportedValue =
+type ISupportedValue =
   | Record<string, unknown>
   | string
   | boolean
   | number
   | null;
 
-export function useDeepCompareMemoize(value: SupportedValue): SupportedValue {
-  const ref = useRef<SupportedValue>(null);
+export function useDeepCompareMemoize(value: ISupportedValue): ISupportedValue {
+  const ref = useRef<ISupportedValue>(null);
   if (!dequal(value, ref.current)) {
     ref.current = value;
   }